split test files and skip flaky tests in linux + collab
