+++ packages/lexical-playground/__tests__/e2e/Links.spec.mjs
@@ -24,6 +24,7 @@ import {
   focusEditor,
   html,
   initialize,
+  IS_LINUX,
   keyDownCtrlOrMeta,
   keyUpCtrlOrMeta,
   test,
@@ -1519,7 +1520,11 @@ test.describe('Links', () => {
     );
   });
 
-  test('Can handle pressing Enter at the end of a Link', async ({page}) => {
+  test('Can handle pressing Enter at the end of a Link', async ({
+    isCollab,
+    page,
+  }) => {
+    test.fixme(isCollab && IS_LINUX, 'Flaky on Linux + Collab');
     await focusEditor(page);
     await page.keyboard.type('Hello awesome');
     await selectAll(page);
